refactor(config): split config by domain
