+++ src/cli-promote.ts
@@ -22,7 +22,7 @@ import {
   resolveFaissIndexBinaryPath,
 } from './active-model.js';
 import { loadManagerForModel, loadWithJsonRetry } from './cli-shared.js';
-import { KNOWLEDGE_BASES_ROOT_DIR } from './config.js';
+import { KNOWLEDGE_BASES_ROOT_DIR } from './config/paths.js';
 import { FaissIndexManager } from './FaissIndexManager.js';
 import { rewriteFileAtomically } from './file-mutation.js';
 import { parseFrontmatter } from './frontmatter.js';
