+++ src/cli-quarantine.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import { KNOWLEDGE_BASES_ROOT_DIR } from './config.js';
+import { KNOWLEDGE_BASES_ROOT_DIR } from './config/paths.js';
 import { listKnowledgeBases } from './kb-fs.js';
 import {
   ackIngestQuarantineEntry,
