+++ src/extraction-cache.ts
@@ -29,7 +29,7 @@
 import { createHash } from 'crypto';
 import * as fsp from 'fs/promises';
 import * as path from 'path';
-import { FAISS_INDEX_PATH } from './config.js';
+import { FAISS_INDEX_PATH } from './config/paths.js';
 import { logger } from './logger.js';
 
 /**
