+++ src/file-ingest.ts
@@ -24,7 +24,8 @@ import * as crypto from 'crypto';
 import * as path from 'path';
 import { Document } from '@langchain/core/documents';
 import { MarkdownTextSplitter, RecursiveCharacterTextSplitter } from 'langchain/text_splitter';
-import { KNOWLEDGE_BASES_ROOT_DIR, resolveChunkSize } from './config.js';
+import { resolveChunkSize } from './config/indexing.js';
+import { KNOWLEDGE_BASES_ROOT_DIR } from './config/paths.js';
 import { handleFsOperationError } from './error-utils.js';
 import { parseFrontmatter } from './frontmatter.js';
 import { detectSiblingPdfPath, liftFrontmatter } from './frontmatter-lift.js';
