+++ src/freshness-manifest.ts
@@ -4,8 +4,10 @@ import * as path from 'path';
 import {
   INGEST_EXCLUDE_PATHS,
   INGEST_EXTRA_EXTENSIONS,
+} from './config/ingest.js';
+import {
   KNOWLEDGE_BASES_ROOT_DIR,
-} from './config.js';
+} from './config/paths.js';
 import { INGEST_BASE_EXTENSIONS } from './ingest-filter.js';
 import { enumerateIngestableKbFiles, listKnowledgeBases } from './kb-fs.js';
