+++ src/lexical-index.ts
@@ -33,7 +33,8 @@ import * as fsp from 'fs/promises';
 import * as path from 'path';
 import { Document } from '@langchain/core/documents';
 import { BM25Retriever } from '@langchain/community/retrievers/bm25';
-import { FAISS_INDEX_PATH, INGEST_EXCLUDE_PATHS, INGEST_EXTRA_EXTENSIONS } from './config.js';
+import { FAISS_INDEX_PATH } from './config/paths.js';
+import { INGEST_EXCLUDE_PATHS, INGEST_EXTRA_EXTENSIONS } from './config/ingest.js';
 import { calculateSHA256, pathExists } from './file-utils.js';
 import { buildChunkDocuments } from './file-ingest.js';
 import { KBError } from './errors.js';
